% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcamr_condum.R
\name{.pcamr_condnum}
\alias{.pcamr_condnum}
\title{Condition number}
\usage{
.pcamr_condnum(M)
}
\arguments{
\item{M}{Numeric matrix.}
}
\value{
Condition number.
}
\description{
Condition number
}
\keyword{internal}
