% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcamr_effective_rank.R
\name{.pcamr_effective_rank}
\alias{.pcamr_effective_rank}
\title{Effective rank of a positive spectrum}
\usage{
.pcamr_effective_rank(vals)
}
\arguments{
\item{vals}{Numeric vector of positive eigenvalues.}
}
\value{
Effective rank.
}
\description{
Effective rank of a positive spectrum
}
\keyword{internal}
