% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcamr_gls_egger.R
\name{.pcamr_gls_egger}
\alias{.pcamr_gls_egger}
\title{GLS Egger in projected PC space}
\usage{
.pcamr_gls_egger(bx_tilde, by_tilde, Omega)
}
\arguments{
\item{bx_tilde}{Projected exposure effects.}

\item{by_tilde}{Projected outcome effects.}

\item{Omega}{Projected outcome covariance matrix.}
}
\value{
Egger fit object.
}
\description{
GLS Egger in projected PC space
}
\keyword{internal}
