% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcamr_validate_inputs.R
\name{.pcamr_validate_inputs}
\alias{.pcamr_validate_inputs}
\title{Validate PCA-MR inputs}
\usage{
.pcamr_validate_inputs(bx, by, bxse, byse, ld)
}
\arguments{
\item{bx}{SNP-exposure effects.}

\item{by}{SNP-outcome effects.}

\item{bxse}{Standard errors of bx.}

\item{byse}{Standard errors of by.}

\item{ld}{LD correlation matrix.}
}
\value{
A validated list of inputs.
}
\description{
Validate PCA-MR inputs
}
\keyword{internal}
