% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcamr_egger.R
\name{pcamr_egger}
\alias{pcamr_egger}
\title{PCA-MR-Egger}
\usage{
pcamr_egger(
  bx,
  by,
  bxse,
  byse,
  ld,
  keep_var = 0.999,
  min_modes = 3,
  kappa_max = 100
)
}
\arguments{
\item{bx}{SNP-exposure effects.}

\item{by}{SNP-outcome effects.}

\item{bxse}{Standard errors of bx.}

\item{byse}{Standard errors of by.}

\item{ld}{LD correlation matrix.}

\item{keep_var}{Cumulative variance explained threshold for PC retention.}

\item{min_modes}{Minimum number of retained PCs. Defaults to 3.}

\item{kappa_max}{Maximum allowed condition number.}
}
\value{
A list containing the PCA-MR-Egger estimate and diagnostics.
}
\description{
PCA-MR-Egger
}
