% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcamr_ivw.R
\name{pcamr_ivw}
\alias{pcamr_ivw}
\title{PCA-MR-IVW}
\usage{
pcamr_ivw(
  bx,
  by,
  bxse,
  byse,
  ld,
  keep_var = 0.999,
  min_modes = 2,
  kappa_max = 100,
  model = c("random", "fixed"),
  phi_floor = 1
)
}
\arguments{
\item{bx}{SNP-exposure effects.}

\item{by}{SNP-outcome effects.}

\item{bxse}{Standard errors of bx.}

\item{byse}{Standard errors of by.}

\item{ld}{LD correlation matrix.}

\item{keep_var}{Cumulative variance explained threshold for PC retention.}

\item{min_modes}{Minimum number of retained PCs.}

\item{kappa_max}{Maximum allowed condition number.}

\item{model}{Either "random" or "fixed".}

\item{phi_floor}{Lower bound for random-effects dispersion.}
}
\value{
A list containing the PCA-MR-IVW estimate and diagnostics.
}
\description{
PCA-MR-IVW
}
