% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcamr_ml.R
\name{pcamr_ml}
\alias{pcamr_ml}
\title{PCA-MR-ML}
\usage{
pcamr_ml(
  bx,
  by,
  bxse,
  byse,
  ld,
  keep_var = 0.999,
  min_modes = 2,
  kappa_max = 100,
  model = c("random", "fixed"),
  alpha = 0.05
)
}
\arguments{
\item{bx}{SNP-exposure effects.}

\item{by}{SNP-outcome effects.}

\item{bxse}{Standard errors of bx.}

\item{byse}{Standard errors of by.}

\item{ld}{LD correlation matrix.}

\item{keep_var}{Cumulative variance explained threshold for PC retention.}

\item{min_modes}{Minimum number of retained PCs.}

\item{kappa_max}{Maximum allowed condition number.}

\item{model}{Either "random" or "fixed".}

\item{alpha}{Confidence interval tail probability.}
}
\value{
A list containing the PCA-MR-ML estimate and diagnostics.
}
\description{
PCA-MR-ML
}
